// generated/src/app/api-models/profsvc-claim-detail-spinal.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProfsvcClaimDetailSpinal {
  seqClaimId: number;
  claimLineNumber: number;
  subLineCode: string;
  spinalLineNumber: number;
  spinalManipConditionCode: string;
  spinalManipConditionDesc1: string;
  spinalManipConditionDesc2: string;
  spinalManipConditionXrayAv: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}